reorg to remove page_controller.ex (which does not make senses if we dont have pages!)
