--- lib/auth_web/templates/page/admin.html.eex
@@ -1,1 +1,0 @@
-<a href="/">Login</a>
