+++ lib/auth_web/views/auth_view.ex
@@ -1,0 +1,3 @@
+defmodule AuthWeb.AuthView do
+  use AuthWeb, :view
+end
