+++ test/auth_web/controllers/auth_controller_test.exs
@@ -2,6 +2,48 @@ defmodule AuthWeb.AuthControllerTest do
   use AuthWeb.ConnCase
   # @email System.get_env("ADMIN_EMAIL")
 
+  test "GET /", %{conn: conn} do
+    conn = get(conn, "/")
+    assert html_response(conn, 200) =~ "Sign in"
+  end
+
+  test "GET /profile (without valid session should redirect)", %{conn: conn} do
+    conn = get(conn, "/profile")
+    # assert html_response(conn, 301) =~ "Login"
+    assert conn.status == 302
+  end
+
+  test "admin/2 show welcome page", %{conn: conn} do
+    data = %{
+      email: "nelson@gmail.com",
+      givenName: "McTestin",
+      picture: "https://youtu.be/naoknj1ebqI",
+      auth_provider: "google"
+    }
+    person = Auth.Person.create_person(data) # |> IO.inspect(label: "person")
+    conn = AuthPlug.create_jwt_session(conn, Map.merge(data, %{id: person.id}))
+    conn = get(conn, "/profile", %{})
+    assert html_response(conn, 200) =~ "google account"
+    # assert html_response(conn, 302) =~ "redirected"
+  end
+
+  test "get_referer/1", %{conn: conn} do
+    conn = conn
+      |> put_req_header("referer", "http://localhost/admin")
+      |> get("/")
+
+    assert conn.resp_body =~ "state=http://localhost/admin"
+  end
+
+  test "get_referer/1 query_string", %{conn: conn} do
+    conn = conn
+      |> get("/?referer=" <> URI.encode("http://localhost/admin")
+          <> "&auth_client_id=" <> AuthPlug.Token.client_id()
+        )
+
+    assert conn.resp_body =~ "state=http://localhost/admin"
+  end
+
   test "github_handler/2 github auth callback", %{conn: conn} do
     conn = get(conn, "/auth/github/callback",
       %{code: "123", state: "http://localhost:4000/" <>
@@ -61,12 +103,21 @@ defmodule AuthWeb.AuthControllerTest do
     assert html_response(conn, 401) =~ "invalid"
   end
 
+  test "login_register_handler/2 with invalid email", %{conn: conn} do
+    conn = post(conn, "/people/register",
+      %{email: "invalid", state: "www.example.com/" <>
+      "&auth_client_id=" <> AuthPlug.Token.client_id() })
+    IO.inspect(conn)
+    # assert html_response(conn, 200) =~ "email"
+    # assert html_response(conn, 401) =~ "invalid"
+  end
+
   test "login_register_handler/2", %{conn: conn} do
     conn = post(conn, "/people/register",
       %{email: "jimmy@dwyl.com", state: "www.example.com/" <>
       "&auth_client_id=" <> AuthPlug.Token.client_id() })
-    IO.inspect(conn)
-    # assert html_response(conn, 200) =~ "google account"
+    # IO.inspect(conn)
+    # assert html_response(conn, 200) =~ "email"
     # assert html_response(conn, 401) =~ "invalid"
   end
 end
